We now have some basic tests for the storage providers
For the build in ones we'll have to manually add the tests for every
single one of them. For the plugin based ones, it should run the
tests on all of the ones that are compiled. Once we added CI it should
compile all the plugins first before running the tests.