Copy the plugin with more permissive permissions
The previous permissions wouldn't allow you to run a second time on the
same working directory as it couldn't delete/overwrite the previous
instance of the plugin when not cleaned up. Like this it can, while I
was at it I also gave execute permission to group, which will make it
easier to do user namespacing in the near future.