Plugins running in a namespace is now optional
Ideally long term it isn't optional, but for now it's just easier to
make it optional so we can still actually run the CI tests properly
(inside docker).

While I was at it, I also noticed that the plugins array in the plugin
manager was never actually filled and therefore processes were never
actually killed properly. So I fixed this and while doing that I also
added a signal handler to the storage server plugin wrapper to cleanly
exit when receiving an interrupt signal.