Implemented support for bigger directories
Changed the interface of the list directory function to return a channel
instead of an array of files. This means we could list directories where
the metadata would be bigger than the system memory (or memory limit of
go-cloud that is). It of course also just means a lower memory usage,
which is also nice. As usual this interface is generic and the lower
memory usage may not hold up for every provider (it doesn't for the
local provider for example).