Initial implementation of a permissions system for plugins
A plugin can specify the network permission in it's manifest, which will
determine whether the namespace/container it will run in will get
network access or not.

As actual namespaced networking seems to be quite a big hack this for
now will piggyback off the host networking (so not namespaced at all) in
case a plugin does want network access. By default however a plugin does
not get any network access.