Added an example of how to have colors in admin panel log
As we write the stdout of the process directly to a buffer checks from
logrus on whether it's a terminal (and thus should use colors) will fail.
This is most likely going to be the case for most logging library like this.
Most have a force option, which works just fine. This just serves as an
example that you should use this force option for this usecase.