Initial implementation of plugin based file info providers
Like storage providers, we also want to be able to run file info
providers as plugins. This updates adds an initial version of this, with
a imagesize as an example, which will display the dimensions of an image
file. There are currently still some performance issues with the http
request for this sometimes stalls for about 20 seconds, which is
something I'm investigating and hopefully fix soon.