This seems to solve fileinfo plugins sometimes getting stuck
Most of this was related to the io.Reader -> grpc -> io.PipeWriter ->
io.Reader bridge, which wasn't sending out EOF correctly. A next step in
this is applying the LimitReader on all plugins. As currently it will
get way more data than needed if any other fileinfo provider wants
everything, which isn't a big deal for the in process ones.. It kind of
is for the plugin ones..

Also I was a dumb dumb and commited generated files.. So I got rid of
those.