Initial prometheus setup
This wraps the storage layer and will provide prometheus metrics for
just about everything going on on the storage layer. It currently
exposes the following metrics related to storage:

init_user
mkdir
move
list_directory
delete

For those it will expose a total time spent and a count. Meaning you can
figure out average time per operation etc. Next to that it will also
expose the following related to files.

open_files
read
write

It should be fairly self explainatory what these are. But for clarity,
open_files is the amount of files currently open. So it's +1 the moment
you use the File method in the StorageProvider. And -1 the moment you
close it. And read and write is the total bytes read or written from/to
the storage provider, which of course can be turned into a data rate
using the prometheus rate() function etc.

Of course this is all optional and disabled by default.