Added a validation step for paths
This function likely will still need work. For now it will just split on
the file seperator and it will check every part on using actually
printable characters. And it explicitly bans the use of .. anywhere, to
not break out to another directory. It also adds a wrapper for the
storage provider which will check all the input before calling the next
and actual storage provider.