Handle file uploads with less memory
Instead of using ParseMultipartForm() we now use the multipart.Reader
directly. Allowing us to read directly from the http body and write it
directly into a storage.File. Next ideal steps would be to implement
some form of javascript uploading, with multiple files support and a
progress bar. Plus resumable file uploads would be a nice to have. I
will likely keep the current way of uploading as a fallback though, as
it works without any form of javascript, but then again should make the
listing of files work without javascript as well in that regard.