Clientside support for uploading multiple files
With this I have also confirmed that uploading without javascript
enabled does in fact work. The only thing that obviously doesn't work is
listing the directory (still need to think how I'll implement this
without javascript). The javascript to handle multiple uploads using tus
is a bit hacky and it uploads them serially, but it works.