Restructured how plugins are launched
Instead of a huge if else statement to switch between local and
namespace mode. It instead now has a global map and a generic interface
that has to be implemented. This means that the whole namespace parts
will only be build on linux (as they should), and won't be used on other
platforms. This will also make it easier to implement similar runners
for other platforms (jails on bsd for example).