Implemented a proxy for prometheus to plugins
It will simply do a /metrics call over the http unix socket of the
plugin. Then parse all the results and return them from the prometheus
on the host process. While adding a plugin label to every result so you
can differentiate between all the different plugins and the host
process. Currently only has support for gauge, counter and summary
types. I still need to add support for the other types and this should
get some tests for sure.