Moved all the current APIs to /webapi/ instead
As these APIs are all specific to the web interface it seems the right
thing to do. My intention is to introduce a more proper API later down
the road. I however don't intend to make the web UI use all of that API
as I intend to make the API mostly json based. And directly streaming
file contents to and from the server. All things that for a web
interface most likely are only practically doable with javascript.