Made debugging storage plugins a bit more user friendly
Now in case a storage plugin doesn't manage to make a grpc connection
due to some error. It will dump the stdout of the plugin into the log
before exiting. It will also attempt to actually stop the plugin
process, in my current test setup however it will keep the unix sockets
around too long. Making the next stop impossible without removing the
unix sockets. I intend to fix that obviously. And I intend to implement
this for the fileinfo plugin system as well (and any future plugin
system).