Changed the buffer size used when transfering large amount of data over unix sockets
This for now is only really relevant on Linux, it will read the
suggested size from /proc/sys/net/core/rmem_default and use this. If
this can't be found for some reason, it will default to 1MB. Which is
actually the case right now when a plugin is running inside a container,
as /proc isn't mounted properly it seems.