Changed the closing of slirp4netns
Previously it would try really hard to end the slirp4netns process,
which probably at that point already exitted by itself as it saw the
network namespace disappear. Trying to Signal(os.Interrupt) and Kill()
the process therefore would always result in an error. So instead we
call Wait() on the process to nicely clean it up (it would go into a
defunct state without this) and just ignore any error from it.

In theory this would mean you could stall the closing of leicht-cloud
forever by nsentering into the network namespace to keep it alive after
the plugin was closed. I would however consider this a non issue, as you
shouldn't be doing that to begin with (besides for debugging I suppose).
And any bad actor being able to do that is a complete different problem
on it's own.