Added basic Apps menu to the navigation bar
While implementing this I realized I would have to pass a list of apps
to every handler, similar to the admin handlers. As this seemed
extremely inefficient and cumbersome, I instead implemented this
completely different. The template handler now has 2 new functions
"apps" and "plugins" which will simply return a list of apps and
plugins. This means it is only stored in one location and it would be a
lot easier to change in the future (when dynamically loading new plugins
or apps becomes possible for example). This also means that the entire
AdminNavbar structure is removed.